+++ evalml/pipelines/time_series_classification_pipelines.py
@@ -7,6 +7,7 @@ from .binary_classification_pipeline_mixin import (
 from evalml.objectives import get_objective
 from evalml.pipelines.classification_pipeline import ClassificationPipeline
 from evalml.pipelines.pipeline_meta import TimeSeriesPipelineBaseMeta
+from evalml.pipelines.time_series_pipeline_base import TimeSeriesPipelineBase
 from evalml.problem_types import ProblemTypes
 from evalml.utils import (
     drop_rows_with_nans,
@@ -15,9 +16,7 @@ from evalml.utils import (
 )
 
 
-class TimeSeriesClassificationPipeline(
-    ClassificationPipeline, metaclass=TimeSeriesPipelineBaseMeta
-):
+class TimeSeriesClassificationPipeline(TimeSeriesPipelineBase, ClassificationPipeline):
     """Pipeline base class for time series classification problems.
 
     Arguments:
@@ -33,28 +32,30 @@ class TimeSeriesClassificationPipeline(
         random_seed (int): Seed for the random number generator. Defaults to 0.
     """
 
-    def __init__(
-        self,
-        component_graph,
-        parameters=None,
-        custom_name=None,
-        random_seed=0,
-    ):
-        if "pipeline" not in parameters:
-            raise ValueError(
-                "date_index, gap, and max_delay parameters cannot be omitted from the parameters dict. "
-                "Please specify them as a dictionary with the key 'pipeline'."
-            )
-        pipeline_params = parameters["pipeline"]
-        self.date_index = pipeline_params["date_index"]
-        self.gap = pipeline_params["gap"]
-        self.max_delay = pipeline_params["max_delay"]
-        super().__init__(
-            component_graph,
-            custom_name=custom_name,
-            parameters=parameters,
-            random_seed=random_seed,
-        )
+    # def __init__(
+    #     self,
+    #     component_graph,
+    #     parameters=None,
+    #     custom_name=None,
+    #     random_seed=0,
+    # ):
+    #     self._encoder = LabelEncoder()
+
+    #     # if "pipeline" not in parameters:
+    #     #     raise ValueError(
+    #     #         "date_index, gap, and max_delay parameters cannot be omitted from the parameters dict. "
+    #     #         "Please specify them as a dictionary with the key 'pipeline'."
+    #     #     )
+    #     # pipeline_params = parameters["pipeline"]
+    #     # self.date_index = pipeline_params["date_index"]
+    #     # self.gap = pipeline_params["gap"]
+    #     # self.max_delay = pipeline_params["max_delay"]
+    #     super().__init__(
+    #         component_graph,
+    #         custom_name=custom_name,
+    #         parameters=parameters,
+    #         random_seed=random_seed,
+    #     )
 
     @staticmethod
     def _convert_to_woodwork(X, y):
@@ -77,25 +78,8 @@ class TimeSeriesClassificationPipeline(
         X, y = self._convert_to_woodwork(X, y)
         self._encoder.fit(y)
         y = self._encode_targets(y)
-
-        self.input_target_name = y.name
-        X_t = self.component_graph.fit_features(X, y)
-
-        y_shifted = y.shift(-self.gap)
-        X_t, y_shifted = drop_rows_with_nans(X_t, y_shifted)
-        self.estimator.fit(X_t, y_shifted)
-        self.input_feature_names = self.component_graph.input_feature_names
+        self._fit(X, y)
         return self
-
-    def _estimator_predict(self, features, y):
-        """Get estimator predictions.
-
-        This helper passes y as an argument if needed by the estimator.
-        """
-        y_arg = None
-        if self.estimator.predict_uses_y:
-            y_arg = y
-        return self.estimator.predict(features, y=y_arg)
 
     def _estimator_predict_proba(self, features, y):
         """Get estimator predicted probabilities.
